requirements.txt
#--------------------------------------------------#
# The following was generated by djinit: #
#--------------------------------------------------#

Django
django-environ
django-jazzmin
djangorestframework
djangorestframework_simplejwt
drf-spectacular
django-cors-headers
whitenoise
gunicorn
dj-database-url
celery
redis
